Moved socket functionality to a separate file.
Also fixed socket usage to test for an uninitialised descriptor.
