+++ src/elk.h
@@ -318,6 +318,10 @@ void entergui();
 /* Plus 1 parallel port */
 void writeparallel(uint8_t val);
 
+/* Socket utilities. */
+int socket_input(int socket_fd, char *buffer, size_t count);
+int socket_open(const char *filename);
+
 #ifdef __cplusplus
 }
 #endif
